Fix bug preventing task desc showing in client
Bug was caused by the previous commit, in which
the task description field was changed from
an input element to a content editable div with
the aim of making the field multi-line and auto
expandable. However, the client code relies on
setting the value of the element, which doesn't
work with a div.

Fix involves changing the task description element
to a textarea.

There is still an issue which needs attention:
new lines in the textarea aren't sent through
to the server, and so aren't saved.