Fix New task values not being saved bug
The issue was that when a new task was added and
the user subsequently entered values into the task
form, the new values weren't updated on the server
(only the new task title was stored). This was
because when a new task was created, the app
created a local copy of the new task data, which
was added to state. When the user made updates to
the values, the local copy was updated, but not
the one stored on the server. I had gone for the
approach of creating a local copy because I was
worried that waiting for the new Task to posted to
the server and then retrieved by the client was
taking too long.

It turns out that my concern about performance was
unfounded; the issue was that I needed to send a response
to the client when the new task had been posted to the
server. Once I added a response in, the performance
issue was no longer a problem.

The server-side code has been generally refactored in
addition to the above, with the dbConnection module now
simply returning a postgres client object rather than
handling the response to the web client. The latter
responsibility has been shifted to the route modules.
This gives greater flexibility and scalability. In
addition, each route now sends a HTTP status code to
the web client in response to a request.