Refactor asynchronous functions
I choose to refactor the getAllTasks & postNewTask
functions so that they use a consistent async/await
syntax rather than a mix of async/await & promise
chaining (in the case of the first), and purely promise
chaining (in the case of the second). This was motivated
by me reading more about how async/await works,
including how it's intended to provide a clearer syntax
than promise chaining.