Add ability to sort tasks by scheduled date
SortBy component now has functionality to sort
tasks by scheduled date (ascending & descending).

Work remains on handling null values, and
adding the option to return to the default
order.

I corrected a bug in the tests for the sort
functionality.

I also made a minor change to the
mockTaskData module to make the syntax cleaner.