Add ability to sort by Task Created date
Can now sort by Task Created Date and Task
Scheduled date, both ascending & descending.

During the feature development, several
unanticipated issues became apparent:

- the onChange event handler attached to the
SortBy select element doesn't fire when the same
value is selected. Solution was to add a button
and sort in response to the button click.

- when the scheduled date is updated locally,
the date is captured, but not the time (the time
is added and set to a default value when the data
is added to the database). This caused an issue
with the sort function when it compares locally
updated dates to dates which hadn't been updated
locally. Solution was to amend the code which
handles task updates so that the time (always a
default value, not the actual time) is added locally
too.

- when the scheduled date is removed locally,
state is set to an emtpy string, whereas in the
database it's set to null. This caused another
conflict between data fetched from the database
and locally modified data. Solution was to amend
the code which handles task updates so that
removing the date locally meant that state was set
to null.

- Some tests needed to be updated to reflect the
above changes,,and to fix some errors made when
they were originally written.

TODO:

- address issue in which sort is restored to default
(Created Date Descending) when deleting a task or
adding a new task. This can result in a poor user
experience.
- improve sort button styling