Stop GET request if adding/deleting causes error
If the server returns an error when the user tries
to add or delete a task, the client will not then
send a subsequent GET request to refresh the task list.

This was causing a clear issue when you try to add
a duplicate task. The server rejects the action but
the GET reuqest still returns the tasks sorted as
though there were a new task.